factories namespace now holds all classes for creating and describing Modules and Calls.
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/core/trunk@701 77d55b08-817d-0410-979b-1fec1beb8c2e

+++ src/cluster/PowerwallView.cpp
@@ -41,8 +41,8 @@ cluster::PowerwallView::~PowerwallView(void) {
  * cluster::PowerwallView::Render
  */
 void cluster::PowerwallView::Render(const mmcRenderViewContext& context) {
-    float time = context.Time;
-    float instTime = context.InstanceTime;
+    float time = static_cast<float>(context.Time);
+    float instTime = static_cast<float>(context.InstanceTime);
     view::CallRenderView *crv = this->getCallRenderView();
 
     this->commPing();
